+++ src/main/java/com/douyinpay/api/payscore/models/ApiModifyAmountForSPRequest.java
@@ -59,7 +59,7 @@ public class ApiModifyAmountForSPRequest {
     // 业务规则："1. 金额：数字，必须≥0（单位：分） 2. 总金额 =（完结付费项目1…+完结付费项目n）-（完结商户优惠项目1…+完结商户优惠项目n） "。
     // 示例：10000。
     @SerializedName("total_amount")
-    private Integer totalAmount;
+    private Long totalAmount;
     // 字段含义：解约原因。
     // 格式规则：string[1, 50]。
     // 示例：用户取消。
@@ -130,11 +130,11 @@ public class ApiModifyAmountForSPRequest {
         this.postDiscounts = postDiscounts;
     }
 
-    public Integer getTotalAmount() {
+    public Long getTotalAmount() {
         return totalAmount;
     }
 
-    public void setTotalAmount(Integer totalAmount) {
+    public void setTotalAmount(Long totalAmount) {
         this.totalAmount = totalAmount;
     }
